New component tests draft
